backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appnope==0.1.4
asgiref==3.6.0
asttokens==2.4.1
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.6.0
bleach==6.1.0
certifi==2024.7.4
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
colorama==0.4.6
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
dataclasses-json==0.5.7
decorator==5.1.1
defusedxml==0.7.1
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
docopt==0.6.2
executing==2.0.1
faiss-cpu==1.8.0.post1
fastjsonschema==2.20.0
frozenlist==1.3.3
greenlet==2.0.2
hyperlink==21.0.0
idna==3.4
importlib_metadata==8.2.0
incremental==22.10.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
JPype1==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
konlpy==0.5.2
langchain==0.0.176
lxml==4.9.2
MarkupSafe==2.1.5
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.7
mecab-python3==1.0.9
mistune==3.0.2
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.6
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.1
parso==0.8.4
pdfkit==1.0.0
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
Pygments==2.18.0
pyOpenSSL==23.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
pyzmq==26.0.3
redis==4.5.4
referencing==0.35.1
regex==2024.7.24
requests==2.29.0
requests-oauthlib==1.3.1
rpds-py==0.19.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.13
sqlparse==0.4.4
stack-data==0.6.3
tenacity==8.2.2
tiktoken==0.7.0
tinycss2==1.3.0
tornado==6.4.1
tqdm==4.65.0
traitlets==5.14.3
tweepy==3.10.0
Twisted==22.10.0
txaio==23.1.1
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
yarl==1.9.2
zipp==3.19.2
zope.interface==6.0
